requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
pillow==11.1.0
opencv-python-headless==4.10.0.84
scikit-image==0.21.0
scipy==1.13.1
google-genai==0.3.0
python-multipart==0.0.20
python-dotenv==1.0.1
